+++ packages/govuk-frontend/src/govuk/macros/i18n.unit.test.mjs
@@ -1,4 +1,4 @@
-import { renderMacro } from 'govuk-frontend-helpers/nunjucks'
+import { renderMacro } from 'govuk-frontend-lib/components'
 
 describe('i18n.njk', () => {
   describe('govukI18nAttributes', () => {
