+++ shared/helpers/package.json
@@ -24,7 +24,6 @@
     "jest-environment-jsdom": "^29.6.2",
     "jest-environment-node-single-context": "^29.1.0",
     "jest-environment-puppeteer": "^9.0.0",
-    "nunjucks": "^3.2.4",
     "outdent": "^0.8.0",
     "puppeteer": "^21.0.1",
     "sass-embedded": "^1.64.2",
