+++ shared/lib/package.json
@@ -17,6 +17,7 @@
     "govuk-frontend-config": "*",
     "js-yaml": "^4.1.0",
     "minimatch": "^9.0.3",
+    "nunjucks": "^3.2.4",
     "slash": "^3.0.0"
   }
 }
